# openapi.json
{
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "Order": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "deliveryWindow": {
            "enum": [
              "morning",
              "afternoon",
              "evening"
            ],
            "title": "Deliverywindow",
            "type": "string"
          },
          "giftMessage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Giftmessage"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/OrderLine"
            },
            "title": "Lines",
            "type": "array"
          },
          "status": {
            "enum": [
              "pending",
              "confirmed"
            ],
            "title": "Status",
            "type": "string"
          },
          "totalCents": {
            "title": "Totalcents",
            "type": "integer"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          },
          "userId": {
            "title": "Userid",
            "type": "string"
          }
        },
        "required": [
          "id",
          "userId",
          "status",
          "totalCents",
          "lines",
          "giftMessage",
          "deliveryWindow",
          "createdAt",
          "updatedAt"
        ],
        "title": "Order",
        "type": "object"
      },
      "OrderItem": {
        "properties": {
          "productId": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Productid",
            "type": "string"
          },
          "quantity": {
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Quantity",
            "type": "integer"
          }
        },
        "required": [
          "productId",
          "quantity"
        ],
        "title": "OrderItem",
        "type": "object"
      },
      "OrderLine": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "priceCents": {
            "title": "Pricecents",
            "type": "integer"
          },
          "productId": {
            "title": "Productid",
            "type": "string"
          },
          "quantity": {
            "title": "Quantity",
            "type": "integer"
          }
        },
        "required": [
          "productId",
          "name",
          "quantity",
          "priceCents"
        ],
        "title": "OrderLine",
        "type": "object"
      },
      "OrderRequest": {
        "properties": {
          "deliveryWindow": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "enum": [
              "morning",
              "afternoon",
              "evening"
            ],
            "title": "Deliverywindow"
          },
          "giftMessage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "maxLength": 200,
            "title": "Giftmessage"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/OrderItem"
            },
            "maxItems": 50,
            "minItems": 1,
            "title": "Items",
            "type": "array"
          },
          "userId": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Userid",
            "type": "string"
          }
        },
        "required": [
          "userId",
          "items"
        ],
        "title": "OrderRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "wmd order-service",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/orders": {
      "get": {
        "operationId": "list_for_user_v1_orders_get",
        "parameters": [
          {
            "in": "query",
            "name": "userId",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 1,
              "title": "Userid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Order"
                  },
                  "title": "Response List For User V1 Orders Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List For User"
      },
      "post": {
        "operationId": "place_v1_orders_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Place"
      }
    },
    "/v1/orders/{order_id}": {
      "get": {
        "operationId": "get_order_v1_orders__order_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Order"
      }
    }
  }
}